+++ ProgramadorGenetico/src/es/si/ProgramadorGenetico/WS/AddProblemaWS.java
@@ -2,8 +2,6 @@ package es.si.ProgramadorGenetico.WS;
 
 import java.net.MalformedURLException;
 import java.net.URL;
-import java.rmi.RemoteException;
-import java.text.DecimalFormat;
 import java.text.NumberFormat;
 import java.util.ArrayList;
 import java.util.Iterator;
@@ -11,8 +9,6 @@ import java.util.List;
 import java.util.Locale;
 
 import javax.xml.namespace.QName;
-
-import org.apache.axis2.AxisFault;
 
 import es.si.ProgramadorGenetico.ServiceWS.Afp;
 import es.si.ProgramadorGenetico.ServiceWS.Configuracion;
@@ -120,18 +116,18 @@ public class AddProblemaWS {
 		this.afp.setEstados(afp.getEstados());
 		
 		String probFinales = "";
-		double[] prob = afp.getProbabilidadesFinal();
+		float[] prob = afp.getProbabilidadesFinal();
 		for (int i = 0; i < estados; i ++)
 			probFinales += (probFinales.equals("") ? prob[i] : ";" + prob[i]);
 		this.afp.setProbFinales(probFinales);
 		
 		String[] trans = new String[estados*2];
-		double[][][] transiciones = afp.getTransiciones();
+		float[][][] transiciones = afp.getTransiciones();
 		for (int i = 0; i < estados; i ++) {
 			for (int j = 0; j < 2; j ++) {
 				trans[i*2 +j] = "" + i + ":" + j + ":";
 				for (int k = 0; k < estados+1; k++) {
-					double temp = (transiciones[i][j][k] < 0.0001 ? 0 : transiciones[i][j][k]);
+					float temp = (transiciones[i][j][k] < 0.0001 ? 0 : transiciones[i][j][k]);
 					temp = (temp > 0.9999 ? 1 : temp);
 					NumberFormat format = NumberFormat.getInstance(Locale.ENGLISH);
 					format.setMaximumFractionDigits(4);
